Bug fixes - Added parentHierarchyChanged() to LevelMeter class, LAF will now be adopted when added to a parent - Changed std::sqrtf to std::sqrt due to GCC bug not implementing it, std::sqrt will have its own overload for float

+++ LevelMeter/LevelMeterSource.h
@@ -99,8 +99,8 @@ private:
         float getAvgRMS () const
         {
             if (rmsHistory.size() > 0)
-                return std::sqrtf (std::accumulate (rmsHistory.begin(), rmsHistory.end(), 0.0f) / rmsHistory.size());
-
+                return std::sqrt(std::accumulate (rmsHistory.begin(), rmsHistory.end(), 0.0f) / static_cast<float>(rmsHistory.size()));
+                
             return float (std::sqrt (rmsSum));
         }
